added function to create generic pointer
