+++ syft/frameworks/torch/pointers/__init__.py
@@ -1,1 +1,5 @@
+from syft.frameworks.torch.pointers.generic_pointer import GenericPointer
+from syft.frameworks.torch.pointers.generic_pointer import create_generic_pointer
 from syft.frameworks.torch.pointers.pointer import *
+
+__all__ = ["GenericPointer", "create_generic_pointer"]
