+++ syft/frameworks/torch/pointers/pointer.py
@@ -1,280 +1,11 @@
 import syft
 import torch
 from syft.frameworks.torch.tensors.interpreters import abstract
-from syft.codes import MSGTYPE
-from syft.exceptions import RemoteObjectFoundError, CannotRequestTensorAttribute
-from typing import List
-from typing import Union
-from typing import TYPE_CHECKING
-
-# this if statement avoids circular imports between base.py and pointer.py
-if TYPE_CHECKING:
-    from syft.workers import BaseWorker
+from syft.exceptions import CannotRequestTensorAttribute
+from syft.frameworks.torch import pointers
 
 
-class GenericPointer(abstract.AbstractObject):
-    """A pointer to a remote object.
-
-    A GenericPointer forwards all API calls to the remote. GenericPointer objects
-    point to objects. They exist to mimic the entire
-    API of a normal tensor, but instead of computing a tensor function locally
-    (such as addition, subtraction, etc.) they forward the computation to a
-    remote machine as specified by self.location. Specifically, every
-    PointerTensor has a tensor located somewhere that it points to (they should
-    never exist by themselves). Note that PointerTensor objects can point to
-    both torch.Tensor objects AND to other PointerTensor objects. Furthermore,
-    the objects being pointed to can be on the same machine or (more commonly)
-    on a different one. Note further that a PointerTensor does not know the
-    nature how it sends messages to the tensor it points to (whether over
-    socket, http, or some other protocol) as that functionality is abstracted
-    in the BaseWorker object in self.location.
-
-    Example:
-
-     >>> import syft as sy
-     >>> hook = sy.TorchHook()
-     >>> bob = sy.VirtualWorker(id="bob")
-     >>> x = sy.Tensor([1,2,3,4,5])
-     >>> y = sy.Tensor([1,1,1,1,1])
-     >>> x_ptr = x.send(bob) # returns a PointerTensor, sends tensor to Bob
-     >>> y_ptr = y.send(bob) # returns a PointerTensor, sends tensor to Bob
-     >>> # executes command on Bob's machine
-     >>> z_ptr = x_ptr + y_ptr
-    """
-
-    def __init__(
-        self,
-        location: "BaseWorker" = None,
-        id_at_location: Union[str, int] = None,
-        owner: "BaseWorker" = None,
-        id: Union[str, int] = None,
-        garbage_collect_data: bool = True,
-        shape: torch.Size = None,
-        point_to_attr: str = None,
-        tags: List[str] = None,
-        description: str = None,
-    ):
-
-        """Initializes a GenericPointer.
-
-        Args:
-            location: An optional BaseWorker object which points to the worker
-                on which this pointer's object can be found.
-            id_at_location: An optional string or integer id of the tensor
-                being pointed to.
-            owner: An optional BaseWorker object to specify the worker on which
-                the pointer is located. It is also where the pointer is
-                registered if register is set to True. Note that this is
-                different from the location parameter that specifies where the
-                pointer points to.
-            id: An optional string or integer id of the PointerTensor.
-            garbage_collect_data: If true (default), delete the remote tensor when the
-                pointer is deleted.
-            point_to_attr: string which can tell a pointer to not point directly to\
-                an object, but to point to an attribute of that object (which must
-                also be a tensor) such as .child or .grad. Note the string can be
-                a chain (i.e., .child.child.child or .grad.child.child). Defaults
-                to None, which meants don't point to any attr, just point to the
-                tensor corresponding to the id_at_location.
-        """
-        super().__init__(id=id, owner=owner, tags=tags, description=description)
-
-        self.location = location
-        self.id_at_location = id_at_location
-        self.garbage_collect_data = garbage_collect_data
-        self._shape = shape
-        self.point_to_attr = point_to_attr
-
-    @classmethod
-    def handle_func_command(cls, command):
-        """
-        Receive an instruction for a function to be applied on a Pointer,
-        Get the remote location to send the command, send it and get a
-        pointer to the response, return.
-        :param command: instruction of a function command: (command name,
-        None, arguments[, kwargs])
-        :return: the response of the function command
-        """
-        pointer = cls.find_a_pointer(command)
-        # Get info on who needs to send where the command
-        owner = pointer.owner
-        location = pointer.location
-
-        # Send the command
-        response = owner.send_command(location, command)
-
-        return response
-
-    @classmethod
-    def find_a_pointer(cls, command):
-        """
-        Find and return the first pointer in the args object, using a trick
-        with the raising error RemoteObjectFoundError
-        """
-        try:
-            cmd, _, args, kwargs = command
-            _ = syft.frameworks.torch.hook_args.hook_function_args(cmd, args, kwargs)
-        except RemoteObjectFoundError as err:
-            pointer = err.pointer
-            return pointer
-
-    def get(self, deregister_ptr: bool = True):
-        raise NotImplementedError("get function of GenericPointer is not implemented")
-
-    def __str__(self):
-        """Returns a string version of this pointer.
-
-        This is primarily for end users to quickly see things about the tensor.
-        This tostring shouldn't be used for anything else though as it's likely
-        to change. (aka, don't try to parse it to extract information. Read the
-        attribute you need directly). Also, don't use this to-string as a
-        serialized form of the pointer.
-        """
-
-        type_name = type(self).__name__
-        out = (
-            f"["
-            f"{type_name} | "
-            f"{str(self.owner.id)}:{self.id}"
-            " -> "
-            f"{str(self.location.id)}:{self.id_at_location}"
-            f"]"
-        )
-
-        if self.point_to_attr is not None:
-            out += "::" + str(self.point_to_attr).replace(".", "::")
-
-        big_str = False
-
-        if self.tags is not None and len(self.tags):
-            big_str = True
-            out += "\n\tTags: "
-            for tag in self.tags:
-                out += str(tag) + " "
-
-        if big_str:
-            out += "\n\tShape: " + str(self.shape)
-
-        if self.description is not None:
-            big_str = True
-            out += "\n\tDescription: " + str(self.description).split("\n")[0] + "..."
-
-        return out
-
-    def __repr__(self):
-        """Returns the to-string method.
-
-        When called using __repr__, most commonly seen when returned as cells
-        in Jupyter notebooks.
-        """
-        return self.__str__()
-
-    def __del__(self):
-        """This method garbage collects the object this pointer is pointing to.
-        By default, PySyft assumes that every object only has one pointer to it.
-        Thus, if the pointer gets garbage collected, we want to automatically
-        garbage collect the object being pointed to.
-        """
-
-        # if .get() gets called on the pointer before this method is called, then
-        # the remote object has already been removed. This results in an error on
-        # this next line because self no longer has .owner. Thus, we need to check
-        # first here and not try to call self.owner.anything if self doesn't have
-        # .owner anymore.
-        if hasattr(self, "owner") and self.garbage_collect_data:
-            # attribute pointers are not in charge of GC
-            if self.point_to_attr is None:
-                self.owner.send_msg(MSGTYPE.FORCE_OBJ_DEL, self.id_at_location, self.location)
-
-    @property
-    def grad(self):
-        if not hasattr(self, "_grad"):
-            self._grad = self.attr("grad")
-
-        if self._grad.child.is_none():
-            return None
-
-        return self._grad
-
-    @grad.setter
-    def grad(self, new_grad):
-        self._grad = new_grad
-
-    @property
-    def data(self):
-        if not hasattr(self, "_data"):
-            self._data = self.attr("data")
-        return self._data
-
-    @data.setter
-    def data(self, new_data):
-        self._data = new_data
-
-    def is_none(self):
-        return self.owner.request_is_remote_tensor_none(self)
-
-    def get_shape(self):
-        """Request information about the shape to the remote worker"""
-        return self.owner.request_remote_tensor_shape(self)
-
-    @property
-    def shape(self):
-        """ This method returns the shape of the data being pointed to.
-        This shape information SHOULD be cached on self._shape, but
-        occasionally this information may not be present. If this is the
-        case, then it requests the shape information from the remote object
-        directly (which is inefficient and should be avoided).
-        """
-
-        if self._shape is None:
-            self._shape = self.get_shape()
-
-        return self._shape
-
-    @shape.setter
-    def shape(self, new_shape):
-        self._shape = new_shape
-
-    def _create_attr_name_string(self, attr_name):
-        if self.point_to_attr is not None:
-            point_to_attr = "{}.{}".format(self.point_to_attr, attr_name)
-        else:
-            point_to_attr = attr_name
-        return point_to_attr
-
-    def attr(self, attr_name):
-        attr_ptr = syft.GenericPointer(
-            id=self.id,
-            owner=self.owner,
-            location=self.location,
-            id_at_location=self.id_at_location,
-            point_to_attr=self._create_attr_name_string(attr_name),
-        )  # .wrap()
-        self.__setattr__(attr_name, attr_ptr)
-        return attr_ptr
-
-    def setattr(self, name, value):
-        self.owner.send_command(
-            message=("__setattr__", self, (name, value), {}), recipient=self.location
-        )
-
-    def share(self, *args, **kwargs):
-        """
-        Send a command to remote worker to additively share a tensor
-
-        Returns:
-            A pointer to an AdditiveSharingTensor
-        """
-
-        # Send the command
-        command = ("share", self, args, kwargs)
-
-        response = self.owner.send_command(self.location, command)
-
-        return response
-
-
-class PointerTensor(GenericPointer, abstract.AbstractTensor):
+class PointerTensor(pointers.GenericPointer, abstract.AbstractTensor):
     """A pointer to another tensor.
 
     A PointerTensor forwards all API calls to the remote.PointerTensor objects
