+++ syft/workers/base.py
@@ -5,7 +5,7 @@ import syft as sy
 
 from syft.frameworks.torch.tensors.interpreters import AbstractTensor
 from syft.generic import ObjectStorage
-from syft.frameworks.torch.pointers import PointerTensor
+from syft.frameworks.torch import pointers
 from syft.exceptions import GetNotPermittedError
 from syft.exceptions import WorkerNotFoundException
 from syft.exceptions import ResponseSignatureError
@@ -262,7 +262,7 @@ class BaseWorker(AbstractWorker, ObjectStorage):
         ptr_id: Union[str, int] = None,
         local_autograd=False,
         preinitialize_grad=False,
-    ) -> PointerTensor:
+    ) -> pointers.PointerTensor:
         """Sends tensor to the worker(s).
 
         Send a syft or torch tensor/object and its child, sub-child, etc (all the
@@ -323,6 +323,17 @@ class BaseWorker(AbstractWorker, ObjectStorage):
                 local_autograd=local_autograd,
                 preinitialize_grad=preinitialize_grad,
             )
+        elif isinstance(obj, torch.jit.ScriptModule):
+            tags = obj.tags if hasattr(obj, "tags") else ""
+            description = obj.description if hasattr(obj, "description") else ""
+            pointer = pointers.create_generic_pointer(
+                owner=self,
+                location=worker,
+                id=ptr_id,
+                id_at_location=obj.id,
+                tags=tags,
+                description=description,
+            )
         else:
             pointer = obj
         # Send the object
@@ -330,7 +341,7 @@ class BaseWorker(AbstractWorker, ObjectStorage):
 
         return pointer
 
-    def execute_command(self, message: tuple) -> PointerTensor:
+    def execute_command(self, message: tuple) -> pointers.PointerTensor:
         """
         Executes commands received from other workers.
 
@@ -395,7 +406,7 @@ class BaseWorker(AbstractWorker, ObjectStorage):
 
     def send_command(
         self, recipient: "BaseWorker", message: str, return_ids: str = None
-    ) -> Union[List[PointerTensor], PointerTensor]:
+    ) -> Union[List[pointers.PointerTensor], pointers.PointerTensor]:
         """
         Sends a command through a message to a recipient worker.
 
@@ -443,7 +454,7 @@ class BaseWorker(AbstractWorker, ObjectStorage):
         # and sent back, so it will be GCed but remote data is any shouldn't be
         # deleted
         if hasattr(obj, "child"):
-            if isinstance(obj.child, PointerTensor):
+            if isinstance(obj.child, pointers.PointerTensor):
                 obj.child.garbage_collect_data = False
             if isinstance(obj.child, (sy.AdditiveSharingTensor, sy.MultiPointerTensor)):
                 shares = obj.child.child
@@ -668,7 +679,7 @@ class BaseWorker(AbstractWorker, ObjectStorage):
     def is_tensor_none(obj):
         return obj is None
 
-    def request_is_remote_tensor_none(self, pointer: PointerTensor):
+    def request_is_remote_tensor_none(self, pointer: pointers.PointerTensor):
         """
         Sends a request to the remote worker that holds the target a pointer if
         the value of the remote tensor is None or not.
@@ -697,7 +708,7 @@ class BaseWorker(AbstractWorker, ObjectStorage):
         """
         return list(tensor.shape)
 
-    def request_remote_tensor_shape(self, pointer: PointerTensor) -> "sy.hook.torch.Size":
+    def request_remote_tensor_shape(self, pointer: pointers.PointerTensor) -> "sy.hook.torch.Size":
         """
         Sends a request to the remote worker that holds the target a pointer to
         have its shape.
@@ -729,7 +740,7 @@ class BaseWorker(AbstractWorker, ObjectStorage):
 
         return None
 
-    def search(self, *query: List[str]) -> List[PointerTensor]:
+    def search(self, *query: List[str]) -> List[pointers.PointerTensor]:
         """Search for a match between the query terms and a tensor's Id, Tag, or Description.
 
         Note that the query is an AND query meaning that every item in the list of strings (query*)
@@ -778,7 +789,7 @@ class BaseWorker(AbstractWorker, ObjectStorage):
 
         return results
 
-    def deserialized_search(self, query_items: Tuple[str]) -> List[PointerTensor]:
+    def deserialized_search(self, query_items: Tuple[str]) -> List[pointers.PointerTensor]:
         """
         Called when a message requesting a call to `search` is received.
         The serialized arguments will arrive as a `tuple` and it needs to be
